+++ NEWS.md
@@ -10,6 +10,7 @@
 
 * All recipe steps now officially support empty selections to be more aligned with dplyr and other packages that use tidyselect (#603, #531). For example, if a previous step removed all of the columns need for a later step, the recipe does not fail when it is estimated (with the exception of `step_mutate()`). The documentation in `?selections` has been updated with advice for writing selectors when filtering steps are used. (#813) 
 
+* `tidy.recipe()` now returns a zero row tibble instead of an error when applied to a empty recipe. (#867)
 
 # recipes 0.1.17